made the survey and the heropage have their own page
