+++ package.json
@@ -14,6 +14,7 @@
     "lodash": "^4.17.21",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.0",
     "react-scripts": "5.0.1",
     "usehooks-ts": "^2.12.1",
     "web-vitals": "^2.1.4"
