+++ src/App.css
@@ -3,3 +3,15 @@ h1 {
   font-size: 96px;
   margin-bottom: 50px;
 }
+.navbar {
+  display: flex;
+  margin-left: auto;
+  gap: 2em;
+  list-style: none;
+  margin-right: 5em;
+  margin-top: 1em;
+}
+.navbarContainer {
+  display: flex;
+  margin-left: auto;
+}
