+++ src/components/HeroComponent/HeroComponent.css
@@ -3,21 +3,25 @@
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  min-height: 100vh;
+  height: auto;
+  margin-top: 7em;
+  margin-bottom: 20em;
 }
 
-.flex-container {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 20px;
+.flexing-container {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 6em;
 }
 
 p {
   flex: 1;
   max-width: 500px;
+  font-family: "Work Sans", sans-serif;
 }
 
 h1 {
   font-size: 96px;
   margin-bottom: 50px;
+  font-family: "Rubik", sans-serif;
 }
