+++ src/components/StudentComponents/StudentComponent.css
@@ -45,7 +45,7 @@ p {
   padding: 20px;
   display: flex;
   flex-direction: column;
-  justify-content: center;
+  align-items: flex-start;
   width: 50%;
 }
 .student-hard-skills {
